+++ Cote/Cote/Domain/Usecase/Note/SearchNoteUseCase.swift
