+++ Cote/Cote/Presentation/View/Component/Sidebar/SearchView.swift
